requirements.txt
pdf2image==1.14.0
PDFNetPython3==8.1.0
Pillow==8.1.0
PyPDF2==1.26.0
